Move some more stuff into control mod.
