+++ src/show.rs
@@ -8,6 +8,7 @@ use crate::{
     channel::{ChannelStateEmitter, Channels},
     clock_service::ClockService,
     config::Config,
+    control::{ControlMessage, Controller},
     dmx::DmxBuffer,
     fixture::{prelude::OscControl, FixtureGroup, FixtureGroupKey, GroupName, Patch},
     master::MasterControls,
@@ -222,53 +223,5 @@ impl Show {
         for group in self.patch.iter() {
             group.render(&self.master_controls, dmx_buffers);
         }
-    }
-}
-
-/// Handle receiving and responding to show control messages.
-struct Controller {
-    osc: OscController,
-    midi: MidiController,
-    recv: Receiver<ControlMessage>,
-}
-
-impl Controller {
-    pub fn from_config(cfg: &Config) -> Result<Self> {
-        let (send, recv) = channel();
-        Ok(Self {
-            osc: OscController::new(cfg.receive_port, cfg.controllers.clone(), send.clone())?,
-            midi: init_midi_controller(&cfg.midi_devices, send)?,
-            recv,
-        })
-    }
-
-    pub fn recv(&self, timeout: Duration) -> Result<Option<ControlMessage>> {
-        match self.recv.recv_timeout(timeout) {
-            Ok(msg) => Ok(Some(msg)),
-            Err(RecvTimeoutError::Timeout) => Ok(None),
-            Err(RecvTimeoutError::Disconnected) => {
-                bail!("control receiver disconnected");
-            }
-        }
-    }
-
-    /// Return a decorated version of self that will include the provided
-    /// metadata when sending OSC response messages.
-    pub fn sender_with_metadata<'a>(
-        &'a self,
-        sender_id: Option<&'a OscClientId>,
-    ) -> OscMessageWithMetadataSender<'_> {
-        self.osc.sender_with_metadata(sender_id)
-    }
-}
-
-pub enum ControlMessage {
-    Osc(OscControlMessage),
-    Midi(MidiControlMessage),
-}
-
-impl CreateControlEvent<Device> for ControlMessage {
-    fn from_event(event: tunnels::midi::Event, device: Device) -> Self {
-        Self::Midi(MidiControlMessage { device, event })
     }
 }
